+++ packages/dynamic-bonding-curve/src/types.ts
@@ -345,25 +345,48 @@ export type CreatePoolParam = {
 }
 
 export type CreateConfigAndPoolParam = CreateConfigParam & {
-    createPoolParam: {
-        name: string
-        symbol: string
-        uri: string
-        poolCreator: PublicKey
-        baseMint: PublicKey
-    }
+    preCreatePoolParam: PreCreatePoolParam
 }
 
 export type CreateConfigAndPoolWithFirstBuyParam = CreateConfigAndPoolParam & {
-    swapBuyParam: {
-        buyAmount: BN
-        minimumAmountOut: BN
-        referralTokenAccount: PublicKey | null
-    }
+    firstBuyParam: FirstBuyParam
 }
 
 export type CreatePoolWithFirstBuyParam = {
     createPoolParam: CreatePoolParam
+    firstBuyParam: FirstBuyParam
+}
+
+export type CreatePoolWithPartnerAndCreatorFirstBuyParam = {
+    createPoolParam: CreatePoolParam
+    partnerFirstBuyParam: PartnerFirstBuyParam
+    creatorFirstBuyParam: CreatorFirstBuyParam
+}
+
+export type PreCreatePoolParam = {
+    name: string
+    symbol: string
+    uri: string
+    poolCreator: PublicKey
+    baseMint: PublicKey
+}
+
+export type FirstBuyParam = {
+    buyer: PublicKey
+    buyAmount: BN
+    minimumAmountOut: BN
+    referralTokenAccount: PublicKey | null
+}
+
+export type PartnerFirstBuyParam = {
+    partner: PublicKey
+    buyAmount: BN
+    minimumAmountOut: BN
+    referralTokenAccount: PublicKey | null
+}
+
+export type CreatorFirstBuyParam = {
+    creator: PublicKey
     buyAmount: BN
     minimumAmountOut: BN
     referralTokenAccount: PublicKey | null
@@ -376,6 +399,7 @@ export type SwapParam = {
     minimumAmountOut: BN
     swapBaseForQuote: boolean
     referralTokenAccount: PublicKey | null
+    payer?: PublicKey
 }
 
 export type SwapQuoteParam = {